Add files via upload
This project is a web-based application that showcases my skills in HTML, CSS and JavaScript. I have utilized Chart.js as a library for creating interactive charts and graphs on the web page. The application is designed to be responsive and adaptive to different screen sizes and devices, ensuring a user-friendly and consistent experience.